Mount CelestialCanvas and fix GLSL precision link errors
Wraps MushafReader (and the loading/error branches) in ReaderShell with
the CelestialCanvas provider, gated reactively on the ambience store
preference. This is the mount point updates.md called out as the
highest-value next step — until now useAmbience() always resolved to
the no-op default context, so the WebGL engine never ran.

First run failed at shader link time: STAR_VERTEX and STAR_FRAGMENT
disagree on default float precision (highp vs mediump) while sharing
uCalm/uTime/uResolution uniforms, which GLSL ES forbids. Fixed by
raising STAR_FRAGMENT to highp rather than lowering the vertex shader,
since the vertex math wraps uTime-derived offsets through fract() every
frame and mediump's ~10-bit mantissa would show up as mote jitter
within minutes.

Verified in headless Chromium at a 390x844 viewport: live WebGL2
context, canvas sized to the real viewport, ~61fps over a 1s sample,
changing framebuffer pixels confirming actual animation, and
AuroraVeil's CSS bloom layers correctly stepping back once the engine
is active. Real touch ripples, reduced-motion/Data Saver paths, and
device performance remain unverified.

Also makes the engine's init failure loggable in dev builds instead of
fully silent, since that silence is why the precision bugs shipped
unnoticed.

+++ src/lib/ambience/shaders.ts
@@ -160,7 +160,11 @@ void main() {
 }`;
 
 export const STAR_FRAGMENT = `#version 300 es
-precision mediump float;
+// highp to match STAR_VERTEX: uTime, uResolution and uCalm are uniforms
+// shared with that stage, and GLSL ES requires a program's stages to agree
+// on a shared uniform's precision. WebGL2 guarantees highp float support in
+// fragment shaders, so this is safe unconditionally.
+precision highp float;
 
 in float vAlpha;
 in float vWarm;
